TestMergeProject(준혁 보완)  : 장바구니-주문완료창 연습1차 저장 (변형되었으니 받지마세요 )

+++ src/main/java/com/example/joinair/dto/ORDER_DETAIL.java
@@ -8,17 +8,17 @@ import java.util.Date;
 @Getter
 @Setter
 public class ORDER_DETAIL {
-    private Long OD_Num; // 주문상세번호    (ORDER_DETAIL의 기본키, 자동1증가생성 시퀀스 만들어줘야 하고 DB에 적용해줘야함 . )
+    private int OD_Num; // 주문상세번호    (ORDER_DETAIL의 기본키, 자동1증가생성 시퀀스 만들어줘야 하고 DB에 적용해줘야함 . )
     private String OD_Pro_Name; // 주문한 상품명    ( DB컬럼 생성 필요함)
     private int OD_Pro_Price; // 주문한 상품명의 단가  (DB컬럼 생성 필요함)
     private int OD_Qua; // 수량             (session에 담긴 Quantity)
-    private double OD_Totalweight; // 총 무게     (session에 담긴 weight)
-    private double OD_Total_Price; // 총 소계금액  (session에 담긴 Subtotal)
+    private int OD_Totalweight; // 총 무게     (session에 담긴 weight)
+    private int OD_Total_Price; // 총 소계금액  (session에 담긴 Subtotal)
 
 
 
 
-    private Long Orders_Num; // 주문번호    (x)
+    private int Orders_Num; // 주문번호    (x)
     private String OD_Location; // 위치             (x)
     private Date OD_Reqtime; // 요청배달시간         (X)
     private String OD_Request; // 배송요청사항   (X)
